updated ExplorerShortcutHelper project to work as a standalone exe, while extracting common logic for reuse
