+++ src/ExplorerShortcutHelper/ExplorerShortcutHelper.csproj
@@ -60,7 +60,8 @@
     <Compile Include="..\CommonAssemblyInfo.cs">
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="Program.cs" />
+    <Compile Include="ExplorerShortcutHelper.cs" />
+    <Compile Include="ExplorerShortcutAdder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RegistryTools.cs" />
   </ItemGroup>
